Revert "[TECH]: Sépare le usecase SIECLE de celui de FREGATA (PIX-11061)"

+++ api/src/prescription/learner-management/domain/usecases/import-organization-learners-from-siecle-format.js
@@ -1,3 +1,4 @@
+import { FileValidationError } from '../../../../../lib/domain/errors.js';
 import { SiecleXmlImportError } from '../errors.js';
 
 import fs from 'fs/promises';
@@ -18,9 +19,11 @@ const ERRORS = {
   INVALID_FILE_EXTENSION: 'INVALID_FILE_EXTENSION',
 };
 
-const importOrganizationLearnersFromSIECLEXMLFormat = async function ({
+const importOrganizationLearnersFromSIECLEFormat = async function ({
   organizationId,
   payload,
+  format,
+  organizationLearnersCsvService,
   organizationLearnerRepository,
   organizationRepository,
   importStorage,
@@ -28,28 +31,39 @@ const importOrganizationLearnersFromSIECLEXMLFormat = async function ({
     unzip: zip.unzip,
     detectEncoding,
   },
+  i18n,
 }) {
   let organizationLearnerData = [];
 
   const organization = await organizationRepository.get(organizationId);
   const path = payload.path;
 
-  const { file: filePath, directory } = await siecleService.unzip(path);
-  const encoding = await siecleService.detectEncoding(filePath);
+  if (format === 'xml') {
+    const { file: filePath, directory } = await siecleService.unzip(path);
+    const encoding = await siecleService.detectEncoding(filePath);
 
-  const filename = await importStorage.sendFile({ filepath: filePath });
+    const filename = await importStorage.sendFile({ filepath: filePath });
 
-  try {
-    if (directory) {
-      await fs.rm(directory, { recursive: true });
+    try {
+      if (directory) {
+        await fs.rm(directory, { recursive: true });
+      }
+      const readableStream = await importStorage.readFile({ filename });
+      const siecleFileStreamer = await SiecleFileStreamer.create(readableStream, encoding);
+      const parser = SiecleParser.create(organization, siecleFileStreamer);
+
+      organizationLearnerData = await parser.parse();
+    } finally {
+      await importStorage.deleteFile({ filename });
     }
-    const readableStream = await importStorage.readFile({ filename });
-    const siecleFileStreamer = await SiecleFileStreamer.create(readableStream, encoding);
-    const parser = SiecleParser.create(organization, siecleFileStreamer);
-
-    organizationLearnerData = await parser.parse();
-  } finally {
-    await importStorage.deleteFile({ filename });
+  } else if (format === 'csv') {
+    organizationLearnerData = await organizationLearnersCsvService.extractOrganizationLearnersInformation(
+      path,
+      organization,
+      i18n,
+    );
+  } else {
+    throw new FileValidationError(ERRORS.INVALID_FILE_EXTENSION, { fileExtension: format });
   }
 
   if (isEmpty(organizationLearnerData)) {
@@ -77,4 +91,4 @@ const importOrganizationLearnersFromSIECLEXMLFormat = async function ({
   });
 };
 
-export { importOrganizationLearnersFromSIECLEXMLFormat };
+export { importOrganizationLearnersFromSIECLEFormat };
